zynqmp mars_xu3: fix issue with wrong flash sizes
In the previous release zynqmp_qspi driver did not read the is-dual
property from devicetree. In the zynqmp_qspi driver, variable is_dual
was set according to status of the MIO pin. Since it changed, it turned out
that with this property set to 1, flash sizes are doubled. This commit fix
this issue